+++ lib/l10n/app_localizations_ru.dart
@@ -2142,4 +2142,28 @@ class SRu extends S {
 
   @override
   String get typeToFilterHint => 'Фильтр...';
+
+  @override
+  String get insertLink => 'Вставить ссылку';
+
+  @override
+  String get linkText => 'Текст';
+
+  @override
+  String get linkHint => 'Гайд';
+
+  @override
+  String get urlLabel => 'URL';
+
+  @override
+  String get urlHint => 'https://example.com';
+
+  @override
+  String get markdownBold => 'Жирный';
+
+  @override
+  String get markdownItalic => 'Курсив';
+
+  @override
+  String get insert => 'Вставить';
 }